+++ src/main/kotlin/io/openfuture/api/config/handler/AuthenticationSuccessHandler.kt
@@ -8,6 +8,7 @@ import io.openfuture.api.util.CookieUtils
 import org.springframework.security.core.Authentication
 import org.springframework.security.oauth2.core.oidc.user.OidcUser
 import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler
+import java.util.*
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
@@ -22,7 +23,8 @@ class AuthenticationSuccessHandler(
         val principal = authentication.principal as OidcUser
         val persistUser = userService.findByGoogleId(principal.subject) ?: userService.save(User(principal.subject))
 
-        val key = persistUser.openKeys.firstOrNull() ?: keyService.generate(persistUser)
+        val key = persistUser.openKeys.firstOrNull { it.enabled && (it.expiredDate == null || Date().after(it.expiredDate)) }
+                ?: keyService.generate(persistUser)
         CookieUtils.add(response, properties.cookieName!!, key.value)
 
         response.sendRedirect("/scaffolds")